% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.R
\name{Calculate}
\alias{Calculate}
\title{calculate the correct rate of the result of sparse.choosekrl()}
\usage{
Calculate(true, results)
}
\arguments{
\item{true}{true k, r, l}

\item{results}{result returned by sparse.choosekrl()}
}
\description{
calculate the correct rate of the result of sparse.choosekrl()
}
