% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer.R
\name{cer}
\alias{cer}
\title{calculate clustering error rate}
\usage{
cer(bires, data)
}
\arguments{
\item{bires}{the return list of label2()}

\item{data}{the return list of get.data()}
}
\value{
the cer in all modes.
}
\description{
calculate clustering error rate
}
