% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseLambda.R
\name{chooseLambda}
\alias{chooseLambda}
\title{choose the best lambda when the range is given}
\usage{
chooseLambda(x, k, r, l, lambda = NULL, method = "L0")
}
\arguments{
\item{x}{a three-dimensional array}

\item{k}{the clusters number of mode 1}

\item{r}{the clusters number of mode 2}

\item{l}{the clusters number of mode 3}

\item{lambda}{a vector of possible lambda, for example: lambda = c(0,50,100,200,300,400,500,600,700,800,900,1000,1100,1200)}

\item{method}{two options: "L0", "L1". Two methods use different penalties, where "L1" indicating Lasso penalty.}
}
\value{
a list \code{lambda} the lambda with lowest BIC;
               \code{BIC} the corresponding BIC for each lambda in the given range;
               \code{nonzeromus} the number clusters with non-zero mean.
}
\description{
choose the best lambda for a certain range.
}
