% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_kmeans.R
\name{cp_kmeans}
\alias{cp_kmeans}
\title{tensor clustering by performing k-means on cp decomposition}
\usage{
cp_kmeans(x, k, r, l, multiplicative = NULL, max.s = NULL)
}
\arguments{
\item{x}{a three-dimensional array}

\item{k}{...}

\item{r}{...}

\item{l}{...}

\item{multiplicative}{the number of components}

\item{max.s}{the max value when selecting s}
}
\value{
a list with only one element: judgeX
}
\description{
tensor clustering by performing k-means on cp decomposition
}
