% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.choosekrl.R
\name{sim.choosekrl}
\alias{sim.choosekrl}
\title{simulation: choose the best k,r,l to do clustering when the ranges are given}
\usage{
sim.choosekrl(n, p, q, k, r, l, error = 1, sim.times = 5,
  method = "L0", mode = "bic", seed = TRUE)
}
\arguments{
\item{n}{the dimension of mode 1}

\item{p}{the dimension of mode 2}

\item{q}{the dimension of mode 3}

\item{k}{true clusters number of mode 1}

\item{r}{true clusters number of mode 2}

\item{l}{true clusters number of mode 3}

\item{error}{noise when producing the data}

\item{sim.times}{simulation times}

\item{method}{two options: "L0", "L1". Two methods use different penalties, where "L1" indicating Lasso penalty.}

\item{mode}{two options: "bic", "crossvalidation".}

\item{seed}{whether set seed to each simulation.}
}
\value{
A list of the krl result in different iteration.
}
\description{
simulation: choose the best k,r,l to do clustering when the ranges are given
}
